Working tests for page 8 - pre-assignment
https://academy.cosmwasm.com/learn/smart-contracts/sending-funds

The issue was that we needed valid and distinct addresses for the owner and sender.

As I don't know of a proper way to generate addresses during the test, I did this:
I did store_code a few times for the same contract on the app, and used those farther away additional addresses. This works for now, but I suppose it could break for tests with more than one contract.